api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["laurens <voogdlaurens@gmail.com>"]
readme = "README.md"
packages = [{include = "api", from = "src"}]

[tool.poetry.dependencies]
python = "~3.10"
click = "^8.1.7"
fastapi = "^0.115.0"
httpx = "^0.27.2"
uvicorn = "^0.30.6"
ariadne = "^0.23.0"
sqlalchemy = "^2.0.35"
strawberry-graphql = "^0.242.0"
asyncpg = "^0.29.0"
graphene = "^3.4"
starlette-graphene = "^0.4.0"
starlette-graphene3 = "^0.6.0"
redis = "^5.1.1"
requests = "^2.32.3"
openai = "^1.52.2"
tiktoken = "^0.8.0"
langchain = "^0.3.4"
langchain-openai = "^0.2.3"
python-dotenv = "^1.0.1"
postgres = "^4.0"
psycopg2-binary = "^2.9.10"
pgvector = "^0.3.5"
langchain-community = "^0.3.3"
unstructured = "^0.16.1"
libmagic = "^1.0"
python-jose = "^3.3.0"
pydantic-settings = "^2.6.0"
google-cloud-storage = "^2.18.2"
cachetools = "^5.5.0"
docker = "^7.1.0"
aioboto3 = "13.2.0"
boto3 = "1.35.36"
aiohttp = "^3.10.10"
stripe = "^11.2.0"
pandas = "^2.2.3"



[tool.poetry.scripts]
api = "api.cli:cli"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.10.0"
bandit = "^1.7.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
